# openapi.yaml
openapi: 3.0.3
info:
  title: Yuki Trader Pro — API Backend V3
  version: "3.0.0"
  description: >
    API REST du backend Yuki Trader Pro (authentification JWT, synchronisation
    cloud multi-appareils, validation Google Play Billing, notifications push
    FCM, supervision). Le moteur d'analyse (SMC, indicateurs techniques,
    pondération dynamique) n'est PAS exposé par cette API : il reste
    entièrement côté client (voir /analysis.js), conformément au principe
    fondamental du cahier des charges V3 ("le moteur d'analyse V2 reste
    inchangé").
servers:
  - url: http://localhost:4000/api
    description: Développement local
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Créer un compte et ouvrir une session (essai gratuit 7 jours)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                deviceId: { type: string }
                deviceLabel: { type: string }
                platform: { type: string, example: android }
      responses:
        "201":
          description: Compte créé
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "400": { description: Requête invalide }
        "409": { description: E-mail déjà utilisé }
  /auth/login:
    post:
      tags: [Auth]
      summary: Se connecter (multi-appareils — chaque appareil obtient son propre refresh token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
                deviceId: { type: string }
                deviceLabel: { type: string }
                platform: { type: string }
      responses:
        "200":
          description: Connexion réussie
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "401": { description: E-mail ou mot de passe incorrect }
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Échanger un refresh token contre un nouveau couple access/refresh (rotation)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
                deviceId: { type: string }
      responses:
        "200":
          description: Nouveau couple de tokens
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuthResult" }
        "401": { description: Refresh token invalide, expiré ou déjà utilisé }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Révoquer un refresh token (déconnexion de cet appareil)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200": { description: Déconnecté }
  /auth/me:
    get:
      tags: [Auth]
      summary: Profil de l'utilisateur connecté, statut d'essai/abonnement, appareils connus
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Profil utilisateur
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MeResult" }
        "401": { description: Authentification requise }
  /sync/state:
    get:
      tags: [Sync]
      summary: Récupérer l'état applicatif synchronisé (journal, positions, signaux, préférences…)
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: État courant (peut être `null` si jamais synchronisé)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SyncStateResult" }
    put:
      tags: [Sync]
      summary: Enregistrer l'état applicatif courant (dernier écrit gagne, version incrémentale)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state]
              properties:
                state: { type: object, description: "Objet opaque — même format que le `state` local du client V2, non interprété par le serveur" }
      responses:
        "200": { description: Enregistré }
        "413": { description: État trop volumineux }
  /sync/devices:
    get:
      tags: [Sync]
      summary: Lister les appareils connus de l'utilisateur
      security: [{ bearerAuth: [] }]
      responses:
        "200": { description: Liste des appareils }
  /sync/devices/{id}:
    delete:
      tags: [Sync]
      summary: Déconnecter un appareil (révoque ses refresh tokens)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200": { description: Appareil déconnecté }
        "404": { description: Appareil introuvable }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        role: { type: string, enum: [free, pro, admin] }
        createdAt: { type: integer, format: int64 }
        trialUntil: { type: integer, format: int64 }
        subscribed: { type: boolean }
    AuthResult:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        accessToken: { type: string, description: "JWT court (15 min par défaut)" }
        refreshToken: { type: string, description: "JWT long (30 jours par défaut), à rotation à chaque usage" }
        expiresIn: { type: integer, description: "Durée de vie de l'access token, en secondes" }
    MeResult:
      type: object
      properties:
        user: { $ref: "#/components/schemas/User" }
        isPro: { type: boolean }
        isTrialActive: { type: boolean }
        devices:
          type: array
          items: { type: object }
    SyncStateResult:
      type: object
      properties:
        state: { type: object, nullable: true }
        updatedAt: { type: integer, format: int64, nullable: true }
        version: { type: integer }
...